+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/index/SandboxContractStore.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.apiserver
+package com.digitalasset.platform.index
 
 import com.daml.ledger.participant.state.index.v2.ContractStore
 import com.digitalasset.daml.lf.data.Ref.Party
